backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
gunicorn==21.2.0
idna==3.7
Levenshtein==0.25.1
mysqlclient==2.2.1
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pip==24.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
rapidfuzz==3.9.6
requests==2.31.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
XlsxWriter==3.2.0
